requirements/dev.txt
# This file was autogenerated by uv via the following command:
#    uv export --format requirements.txt --no-annotate --no-editable --no-hashes --only-group dev -o requirements/dev.txt
annotated-types==0.7.0
bandit==1.8.6 ; python_full_version < '3.10'
bandit==1.9.4 ; python_full_version >= '3.10'
colorama==0.4.6 ; sys_platform == 'win32'
coverage==7.10.7 ; python_full_version < '3.10'
coverage==7.13.4 ; python_full_version >= '3.10'
exceptiongroup==1.3.1 ; python_full_version < '3.11'
iniconfig==2.1.0 ; python_full_version < '3.10'
iniconfig==2.3.0 ; python_full_version >= '3.10'
librt==0.8.1 ; platform_python_implementation != 'PyPy'
lxml==6.0.2
lxml-stubs==0.5.1
markdown-it-py==3.0.0 ; python_full_version < '3.10'
markdown-it-py==4.0.0 ; python_full_version >= '3.10'
mdurl==0.1.2
mypy==1.19.1
mypy-extensions==1.1.0
packaging==26.0
pathspec==1.0.4
pillow==11.3.0 ; python_full_version < '3.10'
pillow==12.1.1 ; python_full_version >= '3.10'
pluggy==1.6.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-extra-types==2.11.0
pygments==2.19.2
pytest==8.4.2 ; python_full_version < '3.10'
pytest==9.0.2 ; python_full_version >= '3.10'
pyyaml==6.0.3
rich==14.3.3
ruamel-yaml==0.19.1
ruff==0.15.5
setuptools==82.0.1
stevedore==5.5.0 ; python_full_version < '3.10'
stevedore==5.7.0 ; python_full_version >= '3.10'
tomli==2.4.0 ; python_full_version <= '3.11'
typing-extensions==4.15.0
typing-inspection==0.4.2
xmldiff==2.7.0
